Use cached properties for performance
In some cases, the cache backend cannot be used because
the data could potentially be stale at the end of the
request. At the same time, sending multiple requests for
each cycle is too expensive.

Cached properties are the middle ground where the value
is cached as long as the active records are passed around